refactor: rename component and using translation direct in the component
